WIP: split trip editor region place UI (checkpoint)
