+++ ClientApps/trip-editor/src/components/useEditorMutationFeedback.ts
@@ -1,0 +1,68 @@
+import { computed, ref, type ComputedRef } from 'vue';
+import { EditorValidationError } from '../api/tripEditorApi';
+
+type FeedbackOptions = {
+  isDirty: ComputedRef<boolean>;
+  isOrdering: ComputedRef<boolean>;
+  isPlaceDraftOpen: ComputedRef<boolean>;
+  isSaving: ComputedRef<boolean>;
+  placeFields: string[];
+  regionFields: string[];
+};
+
+/// Tracks mutation feedback shared by region and place editor forms.
+export function useEditorMutationFeedback(options: FeedbackOptions) {
+  const saveError = ref<string | null>(null);
+  const validationErrors = ref<Record<string, string[]>>({});
+  const lastSavedAt = ref<string | null>(null);
+
+  const statusText = computed(() => {
+    if (options.isSaving.value) {
+      return 'Saving...';
+    }
+
+    if (options.isOrdering.value) {
+      return 'Saving order...';
+    }
+
+    if (saveError.value) {
+      return 'Save failed';
+    }
+
+    if (options.isDirty.value) {
+      return 'Unsaved changes';
+    }
+
+    return lastSavedAt.value ? `Saved ${lastSavedAt.value}` : 'Saved';
+  });
+
+  const formSummaryErrors = computed(() => {
+    const fields = options.isPlaceDraftOpen.value ? options.placeFields : options.regionFields;
+    return Object.entries(validationErrors.value).filter(([key]) => !fields.includes(key)).flatMap(([, messages]) => messages);
+  });
+
+  function applyError(error: unknown, fallback: string): void {
+    if (error instanceof EditorValidationError) {
+      validationErrors.value = error.errors;
+      saveError.value = error.message;
+      return;
+    }
+
+    saveError.value = error instanceof Error ? error.message : fallback;
+  }
+
+  function fieldErrors(key: string): string[] {
+    return validationErrors.value[key] ?? [];
+  }
+
+  function markSaved(): void {
+    lastSavedAt.value = new Intl.DateTimeFormat(undefined, { timeStyle: 'short' }).format(new Date());
+  }
+
+  function resetFeedback(): void {
+    saveError.value = null;
+    validationErrors.value = {};
+  }
+
+  return { applyError, fieldErrors, formSummaryErrors, markSaved, resetFeedback, saveError, statusText };
+}
